+++ msbuild/Xamarin.iOS.Tasks/Tasks/CompileEntitlements.cs
@@ -1,6 +1,6 @@
 ﻿namespace Xamarin.iOS.Tasks
 {
-	public class CompileEntitlements : CompileEntitlementsTaskBase
+	public class CompileEntitlements : CompileEntitlementsTaskCore
 	{
 	}
 }
